backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
httpx==0.25.0
pydantic==2.4.2
openai==1.3.0
anthropic==0.5.0
google-generativeai==0.3.0
python-decouple==3.8